requirements.txt
asgiref==3.4.1
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.5
configparser==3.5.0
defusedxml==0.7.1
Django==3.2.7
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
ez-setup==0.9
httpie==2.5.0
idna==2.6
mysqlclient==2.0.3
pkg_resources==0.0.0
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.2.4
typing-extensions==3.10.0.2
urllib3==1.22
